pyproject.toml
[tool.poetry]
name = "spotify-playlist-updater"
version = "0.1.0"
description = "Job to automatically update by-artists playlists when new music is released"
authors = ["Clément Sicard <clement@sicard.ch>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
spotipy = "*"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.3"
isort = "^5.12.0"
interrogate = "^1.7.0"

# 1.7.1 is NOT compatible with ARM-based Macs
mypy = "1.7.0"

pre-commit = "^3.6.2"
pytest = "^8.2.0"
poetry-plugin-export = "^1.6.0"
notebook = "^7.2.1"
setuptools = "^71.0.4"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-module = true
fail-under = 100
verbose = 2


[tool.ruff]
line-length = 120
show-source = true


[tool.isort]
py_version = "311"
line_length = 1000
lines_after_imports = 2
single_line_exclusions = ["collections.abc", "typing", "typing_extensions"]
known_typing = ["typing", "types", "typing_extensions"]
sections = [
    "FUTURE",
    "STDLIB",
    "TYPING",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
include_trailing_comma = true
profile = "google"
multi_line_output = 3
indent = 4
color_output = true
sort_order = "native"

[tool.black]
line-length = 120
color = true

[tool.mypy]
strict = true
ignore_missing_imports = true
warn_unused_ignores = false
files = "**/*.py"

[tool.pytest.ini_options]
testpaths = ["./tests/"]
# tests should fail if there is any warning => except the following nones:
filterwarnings = ["error"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
